Fix rebase compatibility: renamed function, moved header, signature mismatch
- _sfpu_round_nearest_int32_ → _sfpu_round_to_nearest_int32_ (renamed upstream)
  in both WH and BH ckernel_sfpu_tanh_derivative.h
- ttnn/operations/creation.hpp → ttnn/operations/creation/creation.hpp (moved upstream)
  in test_tanh_bw_ulp.cpp and test_tanh_fw_ulp.cpp
- Fix tanh_bw declaration/definition signature mismatch: input_grad parameter
  changed from std::optional<Tensor> (by value) to const std::optional<Tensor>&
  (by const reference) in unary_backward.hpp to match unary_backward.cpp
