+++ tests/ttnn/unit_tests/gtests/test_tanh_fw_ulp.cpp
@@ -40,7 +40,7 @@
 #include <tt-metalium/bfloat16.hpp>
 #include <tt-metalium/constants.hpp>
 #include "ttnn/operations/eltwise/unary/unary.hpp"
-#include "ttnn/operations/creation.hpp"
+#include "ttnn/operations/creation/creation.hpp"
 #include "ttnn/operations/core/core.hpp"
 #include "ttnn/tensor/tensor.hpp"
 #include "ttnn/types.hpp"
