+++ ttnn/cpp/ttnn/operations/eltwise/unary_backward/unary_backward.hpp
@@ -941,7 +941,7 @@ std::vector<std::optional<Tensor>> tanh_bw(
     const Tensor& grad_tensor_arg,
     const Tensor& input_tensor_arg,
     const std::optional<MemoryConfig>& output_mem_config = std::nullopt,
-    std::optional<Tensor> input_grad = std::nullopt);
+    const std::optional<Tensor>& input_grad = std::nullopt);
 
 std::vector<std::optional<Tensor>> sqrt_bw(
     const Tensor& grad_tensor_arg,
